Changing uinnt to uint64_t

+++ hexl/include/hexl/thread-pool/thread-pool-vars.hpp
@@ -7,11 +7,11 @@ namespace intel {
 namespace hexl {
 
 /// @brief Number of threads for thread pool
-extern uint HEXL_NUM_THREADS;
+extern uint64_t HEXL_NUM_THREADS;
 /// @brief Number of recursive levels as parallel calls
-extern uint HEXL_NTT_PARALLEL_DEPTH;
+extern uint64_t HEXL_NTT_PARALLEL_DEPTH;
 /// @brief Wait time (ms) before thread enters sleep mode
-extern uint16_t HEXL_THREAD_WAIT_TIME;
+extern uint64_t HEXL_THREAD_WAIT_TIME;
 
 }  // namespace hexl
 }  // namespace intel
