+++ hexl/thread-pool/thread-pool-vars.cpp
@@ -6,14 +6,14 @@
 namespace intel {
 namespace hexl {
 
-uint16_t HEXL_THREAD_WAIT_TIME = 20;  // ~ 100x wake up time on ICX machine
+uint64_t HEXL_THREAD_WAIT_TIME = 20;  // ~ 100x wake up time on ICX machine
 
 #ifdef HEXL_MULTI_THREADING
-uint HEXL_NUM_THREADS = setup_num_threads("HEXL_NUM_THREADS");
-uint HEXL_NTT_PARALLEL_DEPTH = setup_ntt_calls("HEXL_NTT_PARALLEL_DEPTH");
+uint64_t HEXL_NUM_THREADS = setup_num_threads("HEXL_NUM_THREADS");
+uint64_t HEXL_NTT_PARALLEL_DEPTH = setup_ntt_calls("HEXL_NTT_PARALLEL_DEPTH");
 #else
-uint HEXL_NUM_THREADS = 1;
-uint HEXL_NTT_PARALLEL_DEPTH = 0;
+uint64_t HEXL_NUM_THREADS = 1;
+uint64_t HEXL_NTT_PARALLEL_DEPTH = 0;
 #endif
 
 }  // namespace hexl
